post polls and tweets/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.10.6
frozenlist==1.3.3
GetOldTweets3==0.0.11
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.7
packaging==23.0
Pillow==9.4.0
pyquery==2.0.0
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sympy==1.11.1
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.2
torch==2.0.0
tornado==6.2
tqdm==4.65.0
transformers==4.27.3
tweepy==4.13.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
yarl==1.8.2
